Mark diagnostics counters volatile for ISR safety

+++ firmware/src/WebSerial.cpp
@@ -21,14 +21,14 @@ void WebSerial::sendStateSnapshot(const PotentiometerManager &pots,
     }
 
     JsonObject diag = doc.createNestedObject("diagnostics");
-    diag["uart_overruns"] = diagnostics.uartOverrunCount;
-    diag["midi_drops"] = diagnostics.midiDropCount;
-    diag["loop_overruns"] = diagnostics.loopOverrunCount;
-    diag["midi_task_overruns"] = diagnostics.midiTaskOverrunCount;
-    diag["loop_max_us"] = diagnostics.maxLoopMicros;
-    diag["loop_last_us"] = diagnostics.lastLoopMicros;
-    diag["midi_isr_max_us"] = diagnostics.maxProcessMidiMicros;
-    diag["midi_isr_last_us"] = diagnostics.lastProcessMidiMicros;
+    diag["uart_overruns"] = static_cast<uint32_t>(diagnostics.uartOverrunCount);
+    diag["midi_drops"] = static_cast<uint32_t>(diagnostics.midiDropCount);
+    diag["loop_overruns"] = static_cast<uint32_t>(diagnostics.loopOverrunCount);
+    diag["midi_task_overruns"] = static_cast<uint32_t>(diagnostics.midiTaskOverrunCount);
+    diag["loop_max_us"] = static_cast<uint32_t>(diagnostics.maxLoopMicros);
+    diag["loop_last_us"] = static_cast<uint32_t>(diagnostics.lastLoopMicros);
+    diag["midi_isr_max_us"] = static_cast<uint32_t>(diagnostics.maxProcessMidiMicros);
+    diag["midi_isr_last_us"] = static_cast<uint32_t>(diagnostics.lastProcessMidiMicros);
 
     String payload;
     serializeJson(doc, payload);
